Cleanup generated header files
This makes generated headers consistent with Intel SDK sgx_edger8r.

The generated header files will contain only the function prototypes.
Marshalling structs, function id enums, and function tables are generated
only in the appropriate c file.

Also add NO_THROW declaration to corelibc/string.h to make it C++ conformant.

Some tests lock down the fields of the generated structs in all_t.c and all_u.c.
Therefore all_u_wrapper.cpp and all_t_wrapper.cpp are introduced to include
the generated source files together with the test files that need access to the
structs.
