+++ test/comprehensive/enc/all_t_wrapper.cpp
@@ -1,0 +1,7 @@
+// Copyright (c) Open Enclave SDK contributors.
+// Licensed under the MIT License.
+
+#include "all_t.c"
+#include "testbasic.cpp"
+#include "testerrno.cpp"
+#include "teststring.cpp"
